Refactored, removed some TODOs, cleaned up + ready to merge
